agm: avoid overwriting the values of aif id from received payload and read from acdb only
For Auto PL, HGH, selecting the TDM port is not mapped one to one.
To select TERTIARY_TDM, the values of lpaif_id should be two and intf_id is zero.
This kind of mismatch is handled in acdb configuration to select the correct TDM port and gsl_get_tagged_data() reads from acdb.
But this is getting overwritten by module_info payload, which doesn't take care of this value mismatch.
So in order to make this params dependent only on acdb, removed the code which is assigning lpaif_id and int_id values from module_info received payload.

Change-Id: I08aa19939661d32f9529b8adbff005239ae4053f
Signed-off-by: Narsinga Rao Chella <nrchella@codeaurora.org>