agm: compress plugin: correct compress capture timestamp
In playback, we have api to get session time in run time,
supported by SPR module in playback graph. There is no
similar api for capture graph. As of now we need to rely
on latest read buffer timestamps.

The capture buffer timestamps are not accurate as playback
session timestamps. There might be constant difference
between them. since, buffer timestamps do not include the
pipeline delay in the ADSP. Assumption is that pipeline
delay should be constant for a given graph.
As of now, we are not using this buffer timestamps. This
change avoids unnecessary error log which says SPR module
is not found in the capture graph

Change-Id: Id74eba118c4d2f533e8bee04adf53a8ed0ee0552