agm: service: fix to update buffer written size for compress use case
Non-blocking Compress playback does not work after sending some
initial 2-3 packets to DSP.
This happens because AGM is not updating actual written size in
case of partial or no write, due to which application is
treating this as successful write and skipping the packet.

Fix to update buffer written size in case of partial or no buffer
written by GSL for non blocking compress use case.

Change-Id: Iaf3f8f861e2d6fc84c6c4a5e4aea497893afba80