ipc: dbus: fix to update buffer written size
Non-blocking Compress playback does not work after sending some
initial 2-3 packets to DSP.HLOS was skipping to send some packets
to SPF, which is tried for one time to write and failed.

This happens because AGM IPC server is not updating actual written size in
case of partial or no write, due to which application is
treating this as successful write and skipping the packet.

Fix to update buffer written size in case of partial or no buffer
written by AGM server.

Change-Id: Ibd5365d829ecdedc1349ebcabe06426748cf1609