agm: ipc: DBus: Add error handling while freeing session.
Free session if allocated by current session only
Incase of error scenario, same session is getting tried to open again
In that case it is not expected to free the session as the session is not
allocated by current usecase. Hence free session is done if session is
allocated in current usecase only.

Change-Id: I5995ff1f097658f1c82fc39f41dacc40d44bfe3b